add deepseek train.py

+++ deepseek-new/data/shakespeare/prepare.py
@@ -1,3 +1,4 @@
+# prepare.py (modified)
 import os
 import requests
 import tiktoken
@@ -24,10 +25,7 @@ print(f"train has {len(train_ids):,} tokens")
 print(f"val has {len(val_ids):,} tokens")
 
 # export to bin files
-train_ids = np.array(train_ids, dtype=np.uint16)
-val_ids = np.array(val_ids, dtype=np.uint16)
+train_ids = np.array(train_ids, dtype=np.int64)  # Changed to int64
+val_ids = np.array(val_ids, dtype=np.int64)    # Changed to int64
 train_ids.tofile(os.path.join(os.path.dirname(__file__), 'train.bin'))
 val_ids.tofile(os.path.join(os.path.dirname(__file__), 'val.bin'))
-
-# train.bin has 301,966 tokens
-# val.bin has 36,059 tokens
